added replies to comments, fixed bugs and made enhancements

+++ src/controller/commentsController.js
@@ -4,53 +4,88 @@ import "https://deno.land/x/dotenv@v3.2.0/load.ts";
 
 export const upvoteComment = async (ctx) => {
   const url = new URL(ctx.request.headers.get("referer"));
-  const user = await readUserModel.getUser(ctx.db, ctx.user);
 
-  user && (await commentsModel.upvoteComment(ctx, ctx.params.commentId));
+  ctx.user && (await commentsModel.upvoteComment(ctx, ctx.params.commentId));
   ctx.redirect = new Response(null, {
     status: 303,
-    headers: { Location: user ? url.pathname : "/login" },
+    headers: { Location: ctx.user ? url.pathname : "/login" },
   });
   return ctx;
 };
 
 export const downvoteComment = async (ctx) => {
   const url = new URL(ctx.request.headers.get("referer"));
-  const user = await readUserModel.getUser(ctx.db, ctx.user);
 
-  user && (await commentsModel.downvoteComment(ctx, ctx.params.commentId));
+  ctx.user && (await commentsModel.downvoteComment(ctx, ctx.params.commentId));
   ctx.redirect = new Response(null, {
     status: 303,
-    headers: { Location: user ? url.pathname : "/login" },
+    headers: { Location: ctx.user ? url.pathname : "/login" },
   });
   return ctx;
 };
 
 export const deletecomment = async (ctx) => {
   const url = new URL(ctx.request.headers.get("referer"));
+  const comment = await commentsModel.getComment(ctx, ctx.params.commentId);
 
-  await commentsModel.deleteComment(ctx, ctx.params.commentId);
-  ctx.redirect = new Response(null, {
-    status: 303,
-    headers: { Location: url.pathname },
-  });
-  return ctx;
+  if (comment.author === ctx.user.username || ctx.user.accountType == "admin") {
+    await commentsModel.deleteComment(ctx, ctx.params.commentId);
+    ctx.redirect = new Response(null, {
+      status: 303,
+      headers: { Location: url.pathname },
+    });
+    return ctx;
+  } else {
+    return ctx;
+  }
 };
 
 export const comment = async (ctx) => {
-  const user = await readUserModel.getUser(ctx.db, ctx.user);
   const formdata = await ctx.request.formData();
   const comment = formdata.get("comment");
-  user &&
+  ctx.user &&
     (await commentsModel.comment(
       ctx,
       ctx.params.newsId,
       comment,
-      user.profilePicture
+      ctx.user.profilePicture
     ));
   ctx.redirect = new Response(null, {
     status: 302,
-    headers: { Location: user ? "/news/" + ctx.params.newsId : "/login" },
+    headers: { Location: ctx.user ? "/news/" + ctx.params.newsId : "/login" },
   });
   return ctx;
+};
+
+export const reply = async (ctx) => {
+  const formdata = await ctx.request.formData();
+  const reply = formdata.get("reply");
+  ctx.user &&
+    (await commentsModel.reply(
+      ctx,
+      ctx.params.commentId,
+      ctx.user.username,
+      reply,
+      ctx.user.profilePicture
+    ));
+  ctx.redirect = new Response(null, {
+    status: 302,
+    headers: { Location: ctx.user ? "/news/" + ctx.params.newsId : "/login" },
+  });
+  return ctx;
+};
+
+export const deleteReply = async (ctx) => {
+  const url = new URL(ctx.request.headers.get("referer"));
+  const reply = await commentsModel.getReply(ctx, ctx.params.replyId);
+  if (reply.author === ctx.user.username || ctx.user.accountType == "admin") {
+    await commentsModel.deleteReply(ctx, ctx.params.replyId);
+    ctx.redirect = new Response(null, {
+      status: 303,
+      headers: { Location: url.pathname },
+    });
+    return ctx;
+  } else {
+    return ctx;
+  }
 };
